update (test for newpac frequency ramping up)

+++ miner.h
@@ -1049,7 +1049,7 @@ extern bool opt_gekko_gsh_detect;
 extern float opt_gekko_gsc_freq;
 extern float opt_gekko_gsd_freq;
 extern float opt_gekko_gse_freq;
-extern int opt_gekko_gsh_freq;
+extern float opt_gekko_gsh_freq;
 extern int opt_gekko_gsh_vcore;
 extern int opt_gekko_start_freq;
 extern int opt_gekko_step_freq;
